ipmiutil: openssl support needs shared library
Fixes:
http://autobuild.buildroot.net/results/0d4/0d4e0ea8fdb233ea8e86f86b6fa406311a0279e2/
http://autobuild.buildroot.net/results/907/907ae0feb9ad2e4a0da4b38290a2b617cb4360a0/

The ipmiutil buildsystem doesn't take dependencies of libcrypto (E.G. zlib)
into consideration, causing link issues when linking statically.

The build system doesn't handle LIBS, so there is no simple way to fix this
outside patching the buildsystem.  As we are very close to the release, lets
just mark it as !static for now.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>