package/matchbox: get rid of intermediate non-package matchbox
Currently, the matchbox entry in the menu is a non-package entry, that
forcibly selects the matchbox-wm package, which is the real matchbox WM.

So, get rid of the current matchbox option, rename the existing
matchbox-wm to simply matchbox, so we have a real package from the
onset.

Since we're re-using the previous option for the WM package, there is no
need for an entry in the legacy menu.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>