package/matchbox: finally add prompts to all sub-packages
We currently do not have that many packages without a prompt, and when
we do, they are limited to system-related packages that can't be
selected or are forcibly selected under various system-specific
conditions.

This is clearly not the case for the matchbox sub-packages, so just give
them a prompt, even when they are forcibly selected by the main matchbox
package, and can't be selected without it.

That will at least tell the user those packages exist and are to be
built.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>