configs: add altera_socdk_defconfig for Altera Cyclone 5 Development Board
This is largely the same as altera_sockit_defconfig.

It uses a fresher Linux and u-boot than SocKit. It also speeds the
serial port up to 115200.

The post-image script is generalized by adding
BR2_ROOTFS_POST_SCRIPT_ARGS and moving it up the altera directory.
Similarly, the readme is moved up and made more generic.

Signed-off-by: Charles Manning <cdhmanning@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>