board/wandboard: add a genimage config
This patch adds a genimage config file and a post-image script for the
Wandboard, to generate a medium image "sdcard.img", ready to be booted.
The image contains the layout explained in the board readme.txt file:
U-Boot, its environment, and an Ext2 rootfs partition.

The defconfig has been slightly changed to enable this feature. Also
lighten the readme file since the config file is documented and simpler.

Tested on a Wandboard Solo.

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
[Arnout: change the name of the tmp dir, and remove it before using
it.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>