package/linux-backports: switch to using the kernel-module infra
linux-backports manually generates its autoconf.h header, which needs a
.config file.

Using a pre- or post-configure hook does not really work, because
generating that file touches the .config, and thus our .stamp files ar
enot longer properly time-ordered and running a subsequent make would
incorrectly believe it has to rebuild linux-backports.

So, even though this is not a fixup proper, we generate that file from
the _KCONFIG_FIXUP_CMDS to ensure the stampt file is created after we
touch .config.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Petr Vorel <petr.vorel@gmail.com>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: Romain Naour <romain.naour@openwide.fr>
Tested-by: Romain Naour <romain.naour@openwide.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>